Fix bug with validating promotions
This post claimed that if the number of extra pieces (due to promotion)
for one side is nonzero, then the other side must have less than 16
pieces:

https://chess.stackexchange.com/a/1483

However, I've encountered positions where this is not the case. Indeed,
a pawn must have moved out of the way for the promotion to occur, but we
can still end up with 16 pieces (say if we end up with doubled pawns on
a file).