Remove pawn and material hashes
The motivation is that we could use them to cache certain kinds of
evaluations, but it ends up using more memory with not a noticable
performance improvement. The flip side is that move generation is faster
when we don't have to constantly update these extra hashes (which is a
very hot portion of the code).