Lazily evaluate makemove when generating `child` moves
This comes at a cost for raw performance in the `perft` test, but
shows a near 20% speedup in some benchmarks of the search algorithm.

This makes sense, since we will likely not be searching every possible
position thanks to many pruning heuristics. So we can save time by not
running makemove for nodes that will never be searched.