Precalculate check information for the position (lazily)
This is borrowed from how Stockfish does it. We can keep track of
information related to checks (such as pinned pieces) in the position
datatype. Then, this can be used to check whether a move will give
check without running the makemove procedure.

This also overlaps with some of the functionality in the
`Position.Analysis` module, so we can reuse this code.

As a benefit, we can also check for pinned pieces in the SEE algorithm.